+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/validator/Validator.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.project.yaml.validator;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,10 +20,7 @@ package com.walmartlabs.concord.project.yaml.validator;
  * =====
  */
 
-import com.walmartlabs.concord.project.yaml.model.YamlCheckpoint;
-import com.walmartlabs.concord.project.yaml.model.YamlProcessDefinition;
-import com.walmartlabs.concord.project.yaml.model.YamlProject;
-import com.walmartlabs.concord.project.yaml.model.YamlStep;
+import com.walmartlabs.concord.project.yaml.model.*;
 
 import java.util.HashMap;
 import java.util.List;
@@ -41,22 +38,43 @@ public class Validator {
         if (def.getSteps() == null) {
             return;
         }
-        validate(ctx, def.getSteps().toList());
+
+        validateFlowSteps(ctx, def.getName(), def.getSteps().toList());
     }
 
     public static void validate(ValidatorContext ctx, YamlProject prj) {
-        if (prj.getFlows() == null) {
-            return;
-        }
+        validateFlows(ctx, prj.getFlows());
+        validateForms(prj.getForms());
 
-        prj.getFlows().forEach((s, yamlSteps) -> validate(ctx, yamlSteps));
     }
 
-    private static void validate(ValidatorContext ctx, List<YamlStep> steps) {
-        if (steps == null) {
+    private static void validateFlows(ValidatorContext ctx, Map<String, List<YamlStep>> flows) {
+        if (flows == null) {
             return;
         }
+
+        flows.forEach((name, steps) -> validateFlowSteps(ctx, name, steps));
+    }
+
+    private static void validateFlowSteps(ValidatorContext ctx, String name, List<YamlStep> steps) {
+        if (steps == null) {
+            throw new IllegalArgumentException("Flow -> " + name + " does not have any step to execute");
+        }
+
         steps.forEach(s -> validate(ctx, s));
+    }
+
+    private static void validateForms(Map<String, List<YamlFormField>> forms) {
+        if (forms == null) {
+            return;
+        }
+        forms.forEach(Validator::validateFormFields);
+    }
+
+    private static void validateFormFields(String name, List<YamlFormField> fields) {
+        if (fields == null) {
+            throw new IllegalArgumentException("Form -> " + name + " does not contain any field");
+        }
     }
 
     @SuppressWarnings("unchecked")